requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
python-dotenv>=1.0.1
pydantic>=2.5.3
sqlalchemy>=2.0.27
aiosqlite>=0.19.0
httpx>=0.26.0
aiohttp>=3.9.3
python-multipart>=0.0.9 